+++ migrations/2020-11-05-152724_activity_remove_user_id/down.sql
@@ -1,2 +1,2 @@
-ALTER TABLE activity ADD COLUMN user_id INTEGER;
+ALTER TABLE activity ADD COLUMN user_id bigint;
 ALTER TABLE activity DROP COLUMN sensitive;
